+++ benchmarks/verus/numpy_triple_part_2/files/polynomial_laguerre_lagint.rs
@@ -8,12 +8,12 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn lagint(c: Vec<f64>, m: usize, k: Vec<f64>, lbnd: f64, scl: f64) -> (result: Vec<f64>)
+fn lagint(c: Vec<f64>, m: u8, k: Vec<f64>, lbnd: f64, scl: f64) -> (result: Vec<f64>)
     requires 
-        k.len() == m,
+        k.len() == m as usize,
         scl != 0.0,
     ensures 
-        result.len() == c.len() + m
+        result.len() == c.len() + m as usize
 // </vc-spec>
 // <vc-code>
 {
