+++ benchmarks/verus/numpy_triple_part_2/files/strings_rsplit.rs
@@ -8,15 +8,15 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn rsplit(a: Vec<String>, sep: String, maxsplit: usize) -> (result: Vec<Vec<String>>)
+fn rsplit(a: Vec<String>, sep: String, maxsplit: u8) -> (result: Vec<Vec<String>>)
     requires 
         sep@.len() > 0,
     ensures
         result.len() == a.len(),
         forall|i: int| 0 <= i < result.len() ==> result[i].len() > 0,
-        maxsplit == 0 ==> forall|i: int| 0 <= i < result.len() ==> 
+        maxsplit as int == 0 ==> forall|i: int| 0 <= i < result.len() ==> 
             result[i].len() == 1 && result[i][0]@ == a[i]@,
-        forall|i: int| 0 <= i < result.len() ==> result[i].len() <= maxsplit + 1,
+        forall|i: int| 0 <= i < result.len() ==> result[i].len() <= maxsplit as int + 1,
         forall|i: int| 0 <= i < result.len() ==> 
             (a[i]@.len() == 0 ==> result[i].len() == 1 && result[i][0]@.len() == 0),
 // </vc-spec>
